refactor(appkit): extract agents tool-dispatch engine into tool-dispatch.ts
Step 4 (final) of splitting the agents plugin. Moves dispatchToolCall +
runSubAgent — the tool-call budget, approval gate, and sub-agent recursion —
into tool-dispatch.ts as free functions over RunState + a ToolDispatchDeps
object. The plugin builds deps via toolDispatchDeps(); the two executeTool
closures call the free function. RunState moves with them. Tests updated to
invoke the free functions (deps built from the plugin's own builder).

agents.ts: 2088 -> 1839 lines (2512 -> 1839 across all four steps). typecheck +
394 agent tests green.
